auto eat and remove viewer form index
