+++ package.json
@@ -12,6 +12,7 @@
         "config": "^3.3.9",
         "minecraft-data": "^3.27.0",
         "mineflayer": "^4.8.1",
+        "mineflayer-auto-eat": "^3.3.1",
         "mineflayer-pathfinder": "^2.4.2",
         "prismarine-viewer": "^1.23.0",
         "vec3": "^0.1.8"
